Update README.md
Rainfall Statistics

Write a program that lets the user enter the total rainfall for each of 12 months into an array of doubles. The program should calculate and display the total rainfall for the year, the average monthly rainfall, and the months with the highest and lowest amounts (you can just display the month's number, for example 1 for January).

Input Validation: Do not accept negative numbers for monthly rainfall figures.